requirements.txt
absl-py==2.3.1
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
audioread==3.1.0
backports.zstd==1.3.0
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
brotli==1.2.0
cachetools==6.2.5
catalogue==2.0.10
category_encoders==2.7.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
comtypes==1.4.15
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
cymem==2.0.13
Cython==3.2.4
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_daq==0.5.0
debugpy==1.8.20
decorator==5.2.1
deepface==0.0.98
deprecation==2.1.0
dtale==3.20.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.128.0
fastjsonschema==2.21.2
filelock==3.20.3
fire==0.7.1
Flask==3.0.3
Flask-Compress==1.23
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.1.0
future==1.0.0
gast==0.4.0
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
gunicorn==24.1.1
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
imbalanced-learn==0.14.1
importlib_metadata==8.7.1
iniconfig==2.3.0
ipykernel==7.2.0
ipython==8.38.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kaleido==1.2.0
keras==2.13.1
kiwisolver==1.4.9
langchain==1.2.7
langchain-core==1.2.7
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.6
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
lightdsa==0.0.3
lightecc==0.0.4
lightgbm==4.6.0
lightphe==0.0.20
llvmlite==0.46.0
logistro==2.0.1
lxml==6.0.2
lz4==4.4.5
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapipe==0.10.9
missingno==0.5.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
mtcnn==1.0.0
murmurhash==1.0.15
namex==0.1.0
narwhals==2.15.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.63.1
numpy==2.2.6
oauthlib==3.3.1
opencv-contrib-python==4.11.0.86
opencv-python==4.13.0.90
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
pandas==2.3.3
parso==0.8.5
patsy==1.0.2
pillow==12.1.0
pkginfo==1.12.1.2
platformdirs==4.5.1
plotly==6.5.2
plotly-resampler==0.11.0
pluggy==1.6.0
pmdarima==2.0.4
polars==1.38.1
polars-runtime-32==1.38.1
pooch==1.8.2
preshed==3.0.12
prompt_toolkit==3.0.52
protobuf==3.20.3
psutil==7.2.1
pure_eval==0.2.3
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycaret==3.3.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.2
pyod==2.0.6
pyparsing==3.3.2
PyPDF2==3.0.1
pypiwin32==223
pyserial==3.5
PySocks==1.7.1
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
pyttsx3==2.99
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retina-face==0.0.17
retrying==1.4.2
rich==14.3.1
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.7.2
scikit-plot==0.3.7
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.1
shap==0.49.1
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sklearn-compat==0.1.5
sktime==0.26.0
slicer==0.0.8
smart_open==7.5.0
smmap==5.0.2
sounddevice==0.5.5
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
squarify==0.4.4
srsly==2.5.2
stack-data==0.6.3
starlette==0.50.0
statsmodels==0.14.6
streamlit==1.53.1
streamlit-autorefresh==1.0.1
streamlit-server-state==0.20.2
strsimpy==0.2.1
sympy==1.14.0
tbats==1.1.3
tenacity==9.1.2
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
tf_keras==2.20.1
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.4.0
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.36.2
tsdownsample==0.1.4.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.12
ultralytics-thop==2.0.18
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.3.5
weasel==0.4.3
Werkzeug==3.0.6
widgetsnbextension==4.0.15
wrapt==2.0.1
xarray==2025.6.1
xlrd==2.0.2
xxhash==3.6.0
yellowbrick==1.5
zipp==3.23.0
zstandard==0.25.0
